+++ src/app/dashboard/components/game-characters-tile/game-characters-tile.component.html
@@ -1,1 +1,17 @@
-<p>game-characters-tile works!</p>
+<div class="tile-box">
+    <ng-container *ngIf="characters.length">
+        <div *ngFor="let item of characters" class="tile">
+            <div class="class">{{item.class}}</div>
+            <div class="class">{{item.race}}</div>
+            <div class="name">{{item.name}}</div>
+        </div>
+    </ng-container>
+    
+    <div class="add-new-charecter tile">
+        <button mat-fab color="primary" (click)="addNewCharecter()" >
+            <mat-icon  aria-label="Add new charecter">add</mat-icon>
+        </button>
+    </div>
+</div>
+
+
