+++ src/app/dashboard/components/user-profile/user-profile.component.html
@@ -41,7 +41,6 @@
                     formControlName="mainCharecters">
           </mat-form-field>
     </form>
-    <hr>
     <!-- <mat-card class="characters" *ngFor="let char of profileData.cherecters">
         <div class="char-container">
             <div class="img-container">
