+++ src/app/dashboard/containers/user-profile-container/user-profile-container.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { user } from '../../../shared/models/constants';
 
 @Component({
   selector: 'app-user-profile-container',
@@ -6,6 +7,21 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./user-profile-container.component.scss']
 })
 export class UserProfileContainerComponent implements OnInit {
+
+  characters = [
+    {
+      name: 'Aizik',
+      race: 'ork',
+      class: 'paladin'
+    },
+    {
+      name: 'Utumba',
+      race: 'ork',
+      class: 'shaman'
+    }
+  ];
+
+  userData = user;
 
   constructor() { }
