+++ src/app/shared/models/profile.model.ts
@@ -1,0 +1,6 @@
+export class Profile {
+    gameEmail: string;     // string   используется для рассылки уведомлений
+    googleAvatarURL: string;
+    charecterAvatarURL: string;  // string  если пустой используется googleAvatarURL
+    mainCharecters: string;
+}
