+++ src/styles/variables.scss
@@ -1,0 +1,9 @@
+
+
+/* Path variables */
+$assetsPath: '/assets/';
+$imagesPath: '#{$assetsPath}images/';
+$fontsPath: '#{$assetsPath}fonts/';
+
+$black: #000;
+$white: #fff;
